# fixtures/flagship/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ora-air API",
    "version": "0.1.0",
    "description": "HTTP API for the ora-air demo app. Public read-only flight search is available without auth; booking, seats, and payment flows are intended for trusted first-party clients."
  },
  "servers": [
    {
      "url": "https://flagship-fixture.example.com"
    }
  ],
  "paths": {
    "/api/search": {
      "get": {
        "summary": "Search available flights",
        "parameters": [
          {
            "name": "origin",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Origin IATA code, e.g. JFK"
          },
          {
            "name": "destination",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Destination IATA code, e.g. LAX"
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" },
            "description": "Departure date in YYYY-MM-DD format"
          },
          {
            "name": "passengers",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 9, "default": 1 },
            "description": "Passenger count"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching flight options",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "flights": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Flight" }
                    }
                  },
                  "required": ["flights"]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "Invalid request payload or parameters",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "e": { "type": "string" }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
      "Flight": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "flightNumber": { "type": "string" },
          "origin": { "type": "string" },
          "destination": { "type": "string" },
          "departureTime": { "type": "string" },
          "arrivalTime": { "type": "string" },
          "durationMinutes": { "type": "integer" },
          "aircraft": { "type": "string" },
          "fares": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "classCode": { "type": "string" },
                "className": { "type": "string" },
                "price": { "type": "number" },
                "perks": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              },
              "required": ["classCode", "className", "price", "perks"]
            }
          }
        },
        "required": [
          "id",
          "flightNumber",
          "origin",
          "destination",
          "departureTime",
          "arrivalTime",
          "durationMinutes",
          "aircraft",
          "fares"
        ]
      },
...
    }
  }
}